@@ -1,7 +1,7 @@
 package com.bakhtiyor.gradients.demo
 
 import android.graphics.drawable.shapes.OvalShape
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
